requirements.txt
attrs==19.3.0
backcall==0.1.0
bleach==3.1.4
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.3.0
colorama==0.4.3
correlation-cuda==0.0.0
cycler==0.10.0
cytoolz==0.10.1
dask==2.13.0
decorator==4.4.1
defusedxml==0.6.0
efficientnet-pytorch==0.6.3
entrypoints==0.3
facenet-pytorch==2.2.7
Flask==1.1.2
idna==2.9
imageio==2.8.0
imgaug==0.4.0
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.3
mistune==0.8.4
mkl-fft
mkl-random==1.1.0
mkl-service==2.3.0
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
notebook==6.0.1
numpy==1.18.1
olefile==0.46
opencv-python==4.2.0.34
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.1
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.3
pycparser==2.19
Pygments==2.5.2
pyparsing==2.4.6
pypng==0.0.20
pyrsistent==0.15.7
python-dateutil==2.8.1
pytorch-gradcam==0.2.1
pytubeX==0.1.1
pytz==2019.3
PyWavelets==1.1.1
pywin32==227
pywinpty==0.5.7
pyzmq==18.1.1
requests==2.23.0
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
Shapely==1.7.0
six==1.14.0
terminado==0.8.3
testpath==0.4.4
toolz==0.10.0
torch==1.4.0
torchvision==0.5.0
tornado==6.0.3
tqdm==4.42.1
traitlets==4.3.3
typing-extensions==3.7.4.2
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
youtube-dl==2020.7.28
zipp==2.2.0
